requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.4.2
rsa==4.0
six==1.14.0
tornado==6.0.4
